Fix expr for BSD. Tested on FreeBSD and Debian
From the 'expr' man page on FreeBSD:

http://www.freebsd.org/cgi/man.cgi?query=expr&apropos=0&sektion=0&manpath=FreeBSD+8.0-RELEASE&format=html

The following examples output the number of characters in variable a.
Again, if a might begin with a hyphen, it is necessary to prevent it from
being interpreted as an option to expr.

If the expr command conforms to IEEE Std 1003.1-2001 (``POSIX.1''),
this is simple:
     expr -- "$a" : ".*"